+++ src/Services/wordpressService.tsx
@@ -96,12 +96,9 @@ export async function getWhitePaperContent(slug: string) {
   return data
 }
 
-export async function getIdeasByPage(page: number) {
-  const off = page <= 1 ? 0 : 7 + 9 * (page - 2)
+export async function getIdeasByPage(offset: number) {
   const { data } = await axios.get(
-    `${publicRuntimeConfig.NEXT_APP_BASE_URL}/wp-json/wp/v2/posts?_embed&per_page=${
-      page <= 1 ? '7' : '9'
-    }&offset=${off}`
+    `${publicRuntimeConfig.NEXT_APP_BASE_URL}/wp-json/wp/v2/posts?_embed&per_page=9&offset=${offset}`
   )
   return data
 }
