--- src/components/CategoryTab.css
@@ -1,40 +1,0 @@
-/* purgecss start ignore */
-.agaetis:hover,
-.evenements:hover {
-  background-color: #ff7f40;
-  border-radius: '.125rem';
-  text-decoration: none;
-}
-
-.securite:hover {
-  background-color: #82919c;
-  border-radius: '.125rem';
-  text-decoration: none;
-}
-
-.service-design:hover {
-  background-color: #df1995;
-  border-radius: '.125rem';
-  text-decoration: none;
-}
-
-.data:hover {
-  background-color: #00b2a9;
-  border-radius: '.125rem';
-  text-decoration: none;
-}
-
-.technologie:hover {
-  background-color: #f1b434;
-  border-radius: '.125rem';
-  text-decoration: none;
-}
-
-.strategie-si:hover,
-.all:hover {
-  background-color: #407ec9;
-  border-radius: '.125rem';
-  text-decoration: none;
-}
-
-/* purgecss end ignore */
