+++ src/components/CategoryTab.tsx
@@ -1,10 +1,10 @@
 import clsx from 'clsx'
 import React from 'react'
+import { slugify } from '../Services/textUtilities'
 
 import { Category } from '../types/IdeasContent'
 
 import Button from './Button'
-import './CategoryTab.css'
 
 interface Props {
   categories: Category[]
@@ -12,46 +12,32 @@ interface Props {
   handleFilterChange: (category: string) => void
 }
 
-function slugify(s: string) {
-  const a = 'àáâäæãåāăąçćčđďèéêëēėęěğǵḧîïíīįìłḿñńǹňôöòóœøōõőṕŕřßśšşșťțûüùúūǘůűųẃẍÿýžźż·/_,:;'
-  const b = 'aaaaaaaaaacccddeeeeeeeegghiiiiiilmnnnnoooooooooprrsssssttuuuuuuuuuwxyyzzz------'
-  const p = new RegExp(a.split('').join('|'), 'g')
-
-  return s
-    .toString()
-    .toLowerCase()
-    .replace(/\s+/g, '-') // Replace spaces with -
-    .replace(p, (c) => b.charAt(a.indexOf(c))) // Replace special characters
-    .replace(/&/g, '-and-') // Replace & with 'and'
-    .replace(/[^\w\-]+/g, '') // Remove all non-word characters
-    .replace(/\-\-+/g, '-') // Replace multiple - with single -
-    .replace(/^-+/, '') // Trim - from start of text
-    .replace(/-+$/, '') // Trim - from end of text
-}
-
 function getBorderColor(category: string, selected: boolean) {
-  if (category === 'Agaetis' || category === 'Evènements') {
-    return clsx('hover:bg-orange orange-border-thin', selected ? 'bg-orange text-white' : 'text-orange')
-  } else if (category === 'Stratégie SI') {
-    return clsx('hover:bg-blue blue-border-thin', selected ? 'bg-blue text-white' : 'text-blue')
-  } else if (category === 'Data') {
-    return clsx('hover:bg-teal teal-border-thin', selected ? 'bg-teal text-white' : 'text-teal')
-  } else if (category === 'Service Design') {
-    return clsx('hover:bg-light-pink pink-border-thin', selected ? 'bg-light-pink text-white' : 'text-light-pink')
-  } else if (category === 'Technologie') {
-    return clsx('hover:bg-yellow yellow-border-thin', selected ? 'bg-yellow text-white' : 'text-yellow')
-  } else if (category === 'Agilité') {
-    return clsx(
-      'hover:bg-light-purple purple-light-border-thin',
-      selected ? 'bg-light-purple text-white' : 'text-light-purple'
-    )
-  } else if (category === 'Business Hacking') {
-    return clsx(
-      'hover:bg-red-lighter red-lighter-border-thin',
-      selected ? 'bg-red-lighter text-white' : 'text-red-lighter'
-    )
+  switch (category) {
+    case 'Agaetis':
+    case 'Evènements':
+      return clsx('hover:bg-orange orange-border-thin', selected ? 'bg-orange text-white' : 'text-orange')
+    case 'Stratégie SI':
+      return clsx('hover:bg-blue blue-border-thin', selected ? 'bg-blue text-white' : 'text-blue')
+    case 'Data':
+      return clsx('hover:bg-teal teal-border-thin', selected ? 'bg-teal text-white' : 'text-teal')
+    case 'Service Design':
+      return clsx('hover:bg-light-pink pink-border-thin', selected ? 'bg-light-pink text-white' : 'text-light-pink')
+    case 'Technologie':
+      return clsx('hover:bg-yellow yellow-border-thin', selected ? 'bg-yellow text-white' : 'text-yellow')
+    case 'Agilité':
+      return clsx(
+        'hover:bg-light-purple purple-light-border-thin',
+        selected ? 'bg-light-purple text-white' : 'text-light-purple'
+      )
+    case 'Business Hacking':
+      return clsx(
+        'hover:bg-red-lighter red-lighter-border-thin',
+        selected ? 'bg-red-lighter text-white' : 'text-red-lighter'
+      )
+    default:
+      return clsx('hover:bg-grey grey-border-thin', selected ? 'bg-grey text-white' : 'text-grey')
   }
-  return clsx('hover:bg-grey grey-border-thin', selected ? 'bg-grey text-white' : 'text-grey')
 }
 
 export default function CategoryTab({ categories, categoryFilter, handleFilterChange }: Props) {
@@ -60,7 +46,7 @@ export default function CategoryTab({ categories, categoryFilter, handleFilterCh
       <div className="text-xs mx-auto flex flex-row flex-wrap justify-center">
         <Button
           className={clsx(
-            'uppercase font-semibold my-2 mx-2 p-2 hover:text-white all rounded-full blue-border-thin',
+            'uppercase font-semibold my-2 mx-2 p-2 hover:text-white hover:bg-blue all rounded-full blue-border-thin',
             categoryFilter === 'All' ? 'bg-blue text-white' : 'blue-border-thin text-blue'
           )}
           onClick={() => handleFilterChange('All')}
@@ -73,7 +59,10 @@ export default function CategoryTab({ categories, categoryFilter, handleFilterCh
             className={clsx(
               'uppercase font-semibold my-2 mx-1 md:mx-2 p-2 rounded-full hover:text-white',
               slugify(category.categoryName),
-              getBorderColor(category.categoryName, categoryFilter === category.categoryName)
+              getBorderColor(
+                category.categoryName,
+                categoryFilter.toLocaleLowerCase() === category.categoryName.toLocaleLowerCase()
+              )
             )}
             onClick={() => handleFilterChange(category.categoryName)}
           >
