+++ src/components/Ideas.css
@@ -22,7 +22,3 @@
 .h-07 {
     height: .7rem;
 }
-
-.invisible {
-    visibility: hidden;
-}
